extracting and versioning logic
