+++ requirements.txt
@@ -1,21 +1,27 @@
 alembic==1.11.1
+annotated-types==0.5.0
 anyio==3.6.2
+astroid==2.15.6
 black==23.3.0
 click==8.1.3
+dill==0.3.7
 fastapi==0.95.1
 greenlet==2.0.2
 h11==0.14.0
 httptools==0.5.0
 idna==3.4
+lazy-object-proxy==1.9.0
 Mako==1.2.4
 MarkupSafe==2.1.2
+mccabe==0.7.0
 mypy-extensions==1.0.0
 packaging==23.1
 pathspec==0.11.1
 platformdirs==3.5.1
 psycopg2==2.9.6
-pydantic==1.10.7
-python-decouple==3.8
+psycopg2-binary==2.9.7
+pydantic==1.10.12
+pydantic_core==2.6.1
 python-dotenv==1.0.0
 PyYAML==6.0
 ruff==0.0.270
@@ -23,8 +29,10 @@ sniffio==1.3.0
 SQLAlchemy==2.0.14
 starlette==0.26.1
 tenacity==8.2.2
-typing_extensions==4.5.0
+tomlkit==0.12.1
+typing_extensions==4.7.1
 uvicorn==0.22.0
 uvloop==0.17.0
 watchfiles==0.19.0
 websockets==11.0.3
+wrapt==1.15.0
