+++ src/main.py
@@ -1,19 +1,15 @@
 """main file"""
 
-from typing import Annotated
+from fastapi import APIRouter, FastAPI
 
-from fastapi import APIRouter, Depends, FastAPI
-from sqlalchemy.orm import Session
-
-from src import deps
-from src.crud.crud_location_data import location_data
-from src.schemas.location import LocationInsert
+from src.api.v1.api import v1_router
+from src.core.config import config
 
 app = FastAPI(title="Location collection and querying", openapi_url="/openapi.json")
-router = APIRouter()
+root_router = APIRouter()
 
 
-@router.get("/", status_code=200)
+@root_router.get("/", status_code=200)
 async def root():
     """root method
 
@@ -25,12 +21,5 @@ async def root():
     return {"message": "Hello World"}
 
 
-@router.post("/", status_code=201)
-async def collect(
-    *, loc_in: LocationInsert, db: Annotated[Session, Depends(deps.get_db)]
-):
-    location_data.create(_db=db, obj_in=loc_in)
-    print(loc_in)
-
-
-app.include_router(router=router)
+app.include_router(router=root_router)
+app.include_router(router=v1_router, prefix=config.API_V1_STR)
